+++ src/lifish/entities/Player.cpp
@@ -35,7 +35,7 @@ Player::Player(const sf::Vector2f& pos, const lif::PlayerInfo& info)
 	_init();
 }
 
-Player::Player(const sf::Vector2f& pos, const unsigned short id)
+Player::Player(const sf::Vector2f& pos, const int id)
 	: lif::Entity(pos)
 	, info(id)
 	, drawProxy(*this)
@@ -89,7 +89,7 @@ void Player::_init() {
 	auto& a_right = animated->addAnimation("walk_right");
 	auto& a_left = animated->addAnimation("walk_left");
 
-	for (unsigned i = 0; i < WALK_N_FRAMES; ++i) {
+	for (int i = 0; i < WALK_N_FRAMES; ++i) {
 		a_down.addFrame(sf::IntRect(i * TILE_SIZE, 0, TILE_SIZE, TILE_SIZE));
 		a_up.addFrame(sf::IntRect(i * TILE_SIZE, TILE_SIZE, TILE_SIZE, TILE_SIZE));
 		a_right.addFrame(sf::IntRect(i * TILE_SIZE, 2 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
@@ -100,13 +100,13 @@ void Player::_init() {
 	auto& a_win = animated->addAnimation("win");
 	auto& a_hurt = animated->addAnimation("hurt");
 
-	for (unsigned i = 0; i < DEATH_N_FRAMES; ++i)
+	for (int i = 0; i < DEATH_N_FRAMES; ++i)
 		a_death.addFrame(sf::IntRect(i * TILE_SIZE, 4 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
 	a_win.addFrame(sf::IntRect(DEATH_N_FRAMES * TILE_SIZE, 4 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
 	a_hurt.addFrame(sf::IntRect((DEATH_N_FRAMES + 1) * TILE_SIZE, 4 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
 
 	auto& a_idle = animated->addAnimation("idle");
-	for (unsigned i = 0; i < IDLE_N_FRAMES; ++i)
+	for (int i = 0; i < IDLE_N_FRAMES; ++i)
 		a_idle.addFrame(sf::IntRect((DEATH_N_FRAMES + 2) * TILE_SIZE, 4 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
 	a_idle.addFrame(sf::IntRect((DEATH_N_FRAMES + 3) * TILE_SIZE, 4 * TILE_SIZE, TILE_SIZE, TILE_SIZE));
